parser/htmlparser/tests/mochitest/test_bug579846.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=579846
-->
<head>
  <title>Test for Bug 579846</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript">
    SimpleTest.waitForExplicitFinish();

    var nextAction;
    function testFrameOnLoad(content)
    {
      nextAction(content);
    }

    nextAction = function(acontent)
    {
      nextAction = function(bcontent)
      {
        ok(parseInt(acontent) < parseInt(bcontent), "Content has changed");
        SimpleTest.finish();
      }

      // Give a chance to update
      window.setTimeout(function() {
        var testFrame = document.getElementById('_test_iframe2');
        testFrame.contentWindow.location.href = 'sub_bug579846.sjs';
      }, 100);
    }

    window.onload = function()
    {
      var testFrame = document.getElementById('_test_iframe1');
      testFrame.contentWindow.location.href = 'sub_bug579846.sjs';
    }
  </script>
</head>
<body>
  <iframe src="" id="_test_iframe1"></iframe>
  <iframe src="" id="_test_iframe2"></iframe>
</body>
</html>
